[IMP] shopinvader: product binding wizard
