+++ shopinvader/__init__.py
@@ -9,3 +9,4 @@ from . import components
 from . import services
 from . import http
 from . import models
+from . import wizards
