+++ shopinvader/models/__init__.py
@@ -4,7 +4,9 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from . import partner
-from . import product
+from . import product_filter
+from . import product_product
+from . import product_template
 from . import product_category
 from . import role
 from . import sale
@@ -16,3 +18,5 @@ from . import url
 from . import account_invoice
 from . import notification
 from . import keychain_account
+from . import shopinvader_product
+from . import shopinvader_variant
