+++ shopinvader/views/product_category_view.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <odoo>
 
-
     <record id="product_category_view_form" model="ir.ui.view">
         <field name="model">product.category</field>
         <field name="inherit_id" ref="product.product_category_form_view"/>
@@ -28,14 +27,12 @@
                             <field name="shopinvader_bind_ids" nolabel="1">
                                 <tree>
                                     <field name="backend_id"/>
-                                    <field name="lang_id"/>
                                     <field name="sync_date"/>
                                     <field name="external_id"/>
                                 </tree>
                                 <form>
                                     <group colspan="4" col="4">
                                         <field name="backend_id"/>
-                                        <field name="lang_id"/>
                                     </group>
                                     <notebook>
                                         <page name="description" string="Description">
