+++ shopinvader/views/product_view.xml
@@ -17,7 +17,6 @@
                     <form>
                         <group name="backend">
                             <field name="backend_id" widget="selection"/>
-                            <field name="lang_id" widget="selection"/>
                             <field name="seo_title"/>
                             <field name="meta_keywords"/>
                             <field name="meta_description"/>
@@ -47,6 +46,7 @@
                                     <field name="id"/>
                                     <field name="default_code"/>
                                     <field name="record_id"/>
+                                    <field name="locomotive_backend_ids" widget="many2many_tags"/>
                                 </tree>
                             </field>
                         </group>
@@ -64,6 +64,7 @@
                     <field name="name"/>
                     <field name="based_on"/>
                     <field name="field_id" attrs="{'invisible': [('based_on', '!=', 'field')],
+
                                                'required': [('based_on', '=', 'field')]}"/>
                     <field name="attribute_id" attrs="{'invisible': [('based_on', '!=', 'attribute')],
                                                    'required': [('based_on', '=', 'attribute')]}"/>
@@ -71,6 +72,19 @@
                     <field name="help" colspan="4" nolabel="1"/>
                 </group>
             </form>
+        </field>
+    </record>
+
+    <record id="product_template_form_view" model="ir.ui.view">
+        <field name="name">product.template.common.form</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_form_view"/>
+        <field name="arch" type="xml">
+            <page name="sales" position="inside">
+                <group name="shopinvader" string="Shopinvader">
+                    <field name="locomotive_backend_ids" widget="many2many_tags"/>
+                </group>
+            </page>
         </field>
     </record>
