+++ shopinvader/views/shopinvader_variant.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2017 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+
+<odoo>
+
+    <record model="ir.ui.view" id="shopinvader_variant_search_view">
+        <field name="name">shopinvader.variant.search (in shopinvader)</field>
+        <field name="model">shopinvader.variant</field>
+        <field name="arch" type="xml">
+            <search string="Products">
+                <field name="name" string="Product"/>
+                <field name="backend_id" string="Shopinvader Backend"/>
+                <separator/>
+                <group expand="0" string="Group By">
+                    <filter string="Shopinvader Backend" name="group_by_backend_id" domain="[]" context="{'group_by':'backend_id'}"/>
+                </group>
+            </search>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="shopinvader_variant_tree_view">
+        <field name="name">shopinvader.variant.tree (in shopinvader)</field>
+        <field name="model">shopinvader.variant</field>
+        <field name="arch" type="xml">
+             <tree string="Product" create="0">
+                <field name="backend_id"/>
+                <field name="default_code"/>
+                <field name="name"/>
+             </tree>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="shopinvader_variant_act_window">
+        <field name="name">Shopinvader Variant</field>
+        <field name="res_model">shopinvader.variant</field>
+        <field name="view_mode">tree,form</field>
+        <field name="domain">[]</field>
+         <field name="view_id" ref="shopinvader_variant_tree_view"/>
+        <field name="context">{'search_default_group_by_backend_id': 1}</field>
+        <field name="groups_id" eval="[(4, ref('connector.group_connector_manager'))]"/>
+    </record>
+
+    <record model="ir.actions.act_window" id="shopinvader_variant_no_group_by_act_window">
+        <field name="name">Shopinvader Variant</field>
+        <field name="res_model">shopinvader.variant</field>
+        <field name="view_mode">tree,form</field>
+        <field name="domain">[]</field>
+         <field name="view_id" ref="shopinvader_variant_tree_view"/>
+        <field name="groups_id" eval="[(4, ref('connector.group_connector_manager'))]"/>
+    </record>
+
+    <record model="ir.ui.menu" id="shopinvader_variant_menu">
+        <field name="name">Shopinvader Variant</field>
+        <field name="parent_id" ref="connector_locomotivecms.menu_locomotive_root"/>
+        <field name="action" ref="shopinvader_variant_act_window"/>
+        <field name="sequence" eval="160"/>
+    </record>
+
+</odoo>
