+++ shopinvader/wizards/__init__.py
@@ -1,0 +1,2 @@
+from . import shopinvader_variant_binding_wizard
+from . import shopinvader_variant_unbinding_wizard
