+++ shopinvader/wizards/shopinvader_variant_binding_wizard.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2017 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+
+<odoo>
+
+    <record model="ir.ui.view" id="shopinvader_variant_binding_wizard_form_view">
+        <field name="name">shopinvader.variant.binding.wizard.form (in shopinvader)</field>
+        <field name="model">shopinvader.variant.binding.wizard</field>
+        <field name="arch" type="xml">
+            <form string="Shopinvader Variant Binding Wizard">
+                <field name='backend_id'/>
+                <field name='product_ids'/>
+                <footer>
+                    <button string="Bind Products" name="bind_products"
+                            type="object" class="oe_highlight"/>
+                    or
+                    <button string="Cancel" class="oe_link" special="cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="shopinvader_variant_binding_wizard_act_window">
+        <field name="name">Shopinvader Variant Binding Wizard</field>
+        <field name="res_model">shopinvader.variant.binding.wizard</field>
+        <field name="view_mode">form</field>
+        <field name="context">{}</field>
+        <field name="target">new</field>
+        <field name="view_id" ref="shopinvader_variant_binding_wizard_form_view"/>
+         <field name="groups_id" eval="[(4, ref('connector.group_connector_manager'))]"/>
+    </record>
+
+</odoo>
